Remove last two questions in PR template
Last fall, we introduced this PR template as a response to incidents
that were caused by developers deploying bugs into production and
leaving those bugs in production for awhile due to inadequate monitoring
of changed behavior.

After several manths of using this PR template across the company, we
have continued to have incidents where develoeprs did not ensure that
their changes were properly deployed

Furthermore, as one of the more prolific PR reviewers at the company, I
have personally found that this template has not improved my reviewing
experience because the content that developers put in it do not help me
understand the change more than before.

The optional question about follow-up items is almost always blank,
"N/A" or removed (I just looked at 10 random recent PRs across repos,
and 10/10 were either blank, removed, or "N/A"). Given how sparse this
is, I think this question should be removed to reduce overhead, and
people can feel free to add follow-up items if they wish.

The required question "How will you know this is working after
deployment?" is also problematic because
1. The answers to this question are generally of poor quality (of the 10
   PRs I just checked, five were empty, removed, or "N/A" and three were
   uninformative ("e2e tests", "logs will go down", "nothing will
   happen")
2. Developers' natural workflow does not involve going back through PRs
   that have already been merged and seeing if the actions to take
   post-deployment were performed.

We use Jira for our issue tracking system. Requring developers to verify
their changes prior to resolivng a ticket gets around the second problem
above. If we enforce the ticket resolution requirement, that is superior
to keeping the third question in this PR template.

Sometimes developers use the third question as a way to describe what
the expected behavior changes are. However, I believe that is the wrong
interpretation of the question. "How does htis affect production?" is not
the same as "How will you know this is working in production?". The
answer to "How does this affect production?" should already be answered
in "What does this PR accomplish?". If not, that is when I ask you to
rewrite your PR description.